Fixes #18717 - Make proxies LB and multi-homed
This adds the concept of 'SmartProxyPools' for Smart Proxies,
a SmartProxyPool can be assigned 1 or many Smart Proxies.
When assigned 1 this could then act as a secondary way of
contacting that Smart Proxy. When a SmartProxyPool is assigned
to 2 or more Smart Proxies this would probably be the
Load Balanced address or VIP for the Smart Proxies.

All client communication to Smart Proxy is now done via the SmartProxyPool.
Hosts are now assigned a SmartProxyPool instead of a Smart Proxy.

Signed-off-by: vagrant <vagrant@centos7-devel.sean.example.com>